feat: add duck redirect to web. add load @ start
the duck redirect will check immediately (thanks to @run-at document-start) if the address includes their defined tab in the ia= parameter. if that's not there, redirect to web. this is to fix them annoyingly having images be their default but still showing normal search results. so if you click something then go back in the browser, it takes you to images instead of the actual search