feat: changed sw. add stackexchange
changed the switch back to if. i did this because stack exchange has a
bunch of urls that the same blocks can be used on so I put them in an
array. to my knowledge, switch case can't handle checking against an
array so an if then else process is the option for now.