blocker: add yt position management
created object to manage the position of yt videos by id. when pulled if there's a previous latest position, it will bring up a snackbar allowing you to jump back. every time a video is paused, it will determine if where you're at is further and update accordingly.
TODO: add a threshold percent where it deletes the data